pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src/", "media/Gulama.png", "README.md", "LICENSE", "CHANGELOG.md"]

[project]
name = "gulama"
version = "0.2.0"
description = "Secure, open-source personal AI agent platform — OpenClaw, but secure."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Santosh", email = "santosh@astrafintechlabs.com" },
]
keywords = ["ai", "agent", "security", "personal-assistant", "llm", "chatbot", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
    "Typing :: Typed",
]

dependencies = [
    # Gateway
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "websockets>=14.0",
    "httpx>=0.28",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",

    # LLM — Universal (100+ providers)
    "litellm>=1.56",

    # Memory & Storage
    "chromadb>=0.6",
    "sentence-transformers>=3.3",

    # Security
    "cryptography>=44.0",
    "pyotp>=2.9",
    "keyring>=25.0",

    # Channels
    "python-telegram-bot>=22.0",

    # CLI & Utilities
    "click>=8.1",
    "rich>=13.9",
    "tomli>=2.2",
    "tomli-w>=1.1",
    "structlog>=24.4",
    "diskcache>=5.6",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "bandit>=1.7",
    "pip-audit>=2.7",
]

# Channel integrations
discord = ["discord.py>=2.4"]
slack = ["slack-sdk>=3.33"]
whatsapp = []  # Uses httpx (already in core deps)
matrix = ["matrix-nio[e2e]>=0.24"]
teams = ["botbuilder-core>=4.15"]

# Skills
browser = ["playwright>=1.48"]
browser-ai = ["playwright>=1.48", "browser-use>=0.1", "langchain-openai>=0.2"]
email = []  # Uses stdlib imaplib/smtplib
calendar = ["google-api-python-client>=2.150", "caldav>=1.4"]
google-workspace = ["google-api-python-client>=2.150", "google-auth>=2.30"]
voice = ["openai-whisper>=20240930", "elevenlabs>=1.0", "gtts>=2.5"]
voice-wake = ["pvporcupine>=3.0", "pyaudio>=0.2"]
image = ["openai>=1.50"]

# Security
saml = ["python3-saml>=1.16"]

# Bundles
all-channels = [
    "discord.py>=2.4",
    "slack-sdk>=3.33",
    "playwright>=1.48",
    "matrix-nio[e2e]>=0.24",
]
all-skills = [
    "playwright>=1.48",
    "google-api-python-client>=2.150",
    "google-auth>=2.30",
    "caldav>=1.4",
    "openai>=1.50",
]
full = [
    "discord.py>=2.4",
    "slack-sdk>=3.33",
    "playwright>=1.48",
    "matrix-nio[e2e]>=0.24",
    "google-api-python-client>=2.150",
    "google-auth>=2.30",
    "caldav>=1.4",
    "openai>=1.50",
    "browser-use>=0.1",
    "langchain-openai>=0.2",
]

# Rust acceleration
rust = ["maturin>=1.7"]

[project.scripts]
gulama = "src.cli.commands:cli"

[project.urls]
Homepage = "https://gulama.ai"
Repository = "https://github.com/san-techie21/gulama-bot"
Documentation = "https://gulama.dev"
Issues = "https://github.com/san-techie21/gulama-bot/issues"
Changelog = "https://github.com/san-techie21/gulama-bot/blob/main/CHANGELOG.md"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM", "TC"]
ignore = [
    "S101",   # Allow assert in tests
    "S104",   # Possible binding to all interfaces — intentional for gateway
    "S108",   # Probable insecure temp file usage — controlled by sandbox
    "S110",   # try-except-pass — used intentionally for resilience
    "S603",   # subprocess without shell=True — used in shell_exec skill
    "S607",   # start-process-with-partial-path — controlled by sandbox
    "S608",   # SQL expression — used in memory store (not user-facing)
    "E501",   # Line too long — handled by formatter
    "SIM105", # Use contextlib.suppress — style preference
    "SIM108", # Use ternary — readability preference
    "B017",   # assertRaises(Exception) — acceptable in tests
    "A002",   # Shadowing builtin — acceptable in controlled contexts
    "N806",   # Variable in function not lowercase — acceptable
    "E741",   # Ambiguous variable name — acceptable in context
    "TC001",  # Type-checking import — causes runtime issues if moved
    "TC002",  # Type-checking import — causes runtime issues if moved
    "TC003",  # Type-checking import — causes runtime issues if moved
    "SIM102", # Collapsible if — readability preference
    "SIM103", # Needless bool — style preference
    "SIM110", # Reimplemented builtin — readability preference
    "B904",   # Raise from — acceptable in import error re-raises
    "S105",   # Hardcoded password string — false positive in test/config
    "S106",   # Hardcoded password func arg — false positive in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S105", "S106", "F841", "B007"]  # Allow test-specific patterns

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]  # Allow assert
